requirements.txt
pandas>=2.0
numpy>=1.24
pytz
PyYAML
joblib
matplotlib
tensorboard
MetaTrader5
gymnasium>=0.29
stable-baselines3>=2.2.1